+++ queries/tsx/rainbow-parens.scm
@@ -1,0 +1,9 @@
+; inherits: typescript
+
+(jsx_element
+  (jsx_opening_element ["<" ">"] @opening)
+  (jsx_closing_element ["<" "/" ">"] @closing)) @container
+
+(jsx_self_closing_element
+  "<" @opening
+  ["/" ">"] @closing) @container