+++ queries/tsx/rainbow-tags.scm
@@ -1,0 +1,5 @@
+; inherits: typescript
+
+(jsx_element
+  open_tag: (jsx_opening_element) @opening
+  close_tag: (jsx_closing_element) @closing) @container